+++ src/statics.h
@@ -12,7 +12,7 @@ public:
 	static std::wstring GetModName() { return L"MotorTownMods"; }
 
 	// Get current mod version
-	static std::wstring GetVersion() { return L"0.1.0"; }
+	static std::wstring GetVersion() { return L"0.40.0-rc1"; }
 
 	// Get webhook URL for external callback
 	static const std::string GetWebhookUrl();